[ALPS03773855] DISP: Fix coverity issue
[Detail]
4247181,4247179,4247177,
4247176,4247173,4147169

MTK-Commit-Id: 1445f2d8069d6afeacd2a09d76b77cda2eedc042

CR-Id: ALPS03773855
Feature: Others

+++ drivers/misc/mediatek/video/mt6765/dispsys/ddp_drv.h
@@ -410,9 +410,10 @@ struct disp_iommu_device *disp_get_iommu_dev(void);
 #define DISP_IOCTL_SET_TPLAY_HANDLE \
 	_IOW(DISP_IOCTL_MAGIC, 200, unsigned int)
 
-int disp_get_ovl_bandwidth(unsigned int in_fps, unsigned int out_fps,
+int disp_get_ovl_bandwidth(unsigned long long in_fps,
+			unsigned long long out_fps,
 			unsigned long long *bandwidth);
-int disp_get_rdma_bandwidth(unsigned int out_fps,
+int disp_get_rdma_bandwidth(unsigned long long out_fps,
 			unsigned long long *bandwidth);
 
 #endif
